requirements4.txt
jedi
tzlocal
watchdog
pandas
numpy
scikit-learn
xgboost
fastapi
uvicorn
pydantic
joblib
requests
shap
taipy
faker
matplotlib
plotly
transformers
torch
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
google-generativeai
streamlit
streamlit-aggrid
